improve planning heuristic and some fixes.

+++ code/subarchitectures/planner.sa/trunk/src/base_planners/prob_downward/search/weighted_evaluator.h
@@ -6,7 +6,7 @@
 class WeightedEvaluator : public ScalarEvaluator {
 private:
     ScalarEvaluator* evaluator;
-    int w;
+    double w;
     int value;
 
 public:
